Renamed command_line.rs  correctly
