+++ auto_gippity/src/helpers/command_line.rs
